Move wait utils into dedicated utils module
